contracts-bedrock: fix test legibility
Adds a new helper library called `EIP1967Helper` that can
get the admin and impl storage slots from a `Proxy` implementing
ERC1967. This is more helpful to use than hardcoded magic values
such as `multisig` because it is not clear who the multisig is
since its value is assigned in a different file. We want to decouple
the value from a magic value and set it to exactly what we want it
to be which is the admin. This will work in all cases no matter what
the admin is since it dynamically pulls the value from storage for
the tests.